ASoC: mediatek: mt8173-max98090: Configure jack as a Headset jack
This driver has correctly mapped jack kcontrols for Headphone and
Headset Mic. However, it is configuring the jack to only care about
Headphone events. The MAX98090 codec used here can detect both
connections, so configure the jack as such.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>
Commit-notes:
It also appears to support SND_JACK_BTN_0. Not sure if that should be
part of this patch.
END