clk: mediatek: mt8173: Convert infracfg clk to simple probe
The MT8173 infracfg clock driver can be built as a module, but it fails
to work then because clk_mt8173_infra_init_early() is not called to
allocate the memory regions the probe() will use.

A previous patch adds CPU mux support to mtk_clk_simple_probe() and
mtk_clk_simple_remove(). With that, everything this clock needs is
handled in simple probe functions, so switch it to the common code.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>