dt-bindings: i2c: cros-ec-i2c-tunnel: Document google,limited-passthrough flag
A previous patch implements a limited passthrough mode for the
cros-ec-i2c-tunnel driver which only supports communicating with a smart
battery, due to design restrictions on older boards. This is enabled by
a new "google,limited-passthrough" flag in the corresponding device-tree
node. Document this new property in the associated schema.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>